port deletion of .svn folders to release script
git-svn-id: http://svn.osgeo.org/osgeo/livedvd/gisvm/trunk@12326 7da46fd8-bd2a-0410-a152-cf9810cf5b10

+++ bin/inchroot_release.sh
@@ -168,9 +168,12 @@ export USER_NAME
 
 # save space on ISO by removing the .svn/ dirs
 #   (or control this in bootstrap.sh by uncommenting the 'svn export' line)
-for DIR in `find /usr/local/share/gisvm | grep '\.svn$'` ; do
-   rm -rf "$DIR"
-done
+NEAR_RC=1
+if [ "$NEAR_RC" -eq 1 ] ; then
+   for DIR in `find /usr/local/share/gisvm | grep '\.svn$'` ; do
+      rm -rf "$DIR"
+   done
+fi
 
 # user shouldn't own outside of /home, but a group can
 chown -R root.staff /usr/local/share/gisvm
